chore: refactor to avoid circular imports harder
