+++ src/tbp/monty/frameworks/config_utils/config_args.py
@@ -53,7 +53,7 @@ from tbp.monty.frameworks.loggers.wandb_handlers import (
 )
 from tbp.monty.frameworks.models.abstract_monty_classes import Monty
 from tbp.monty.frameworks.models.displacement_matching import DisplacementGraphLM
-from tbp.monty.frameworks.models.evidence_matching.evidence_matching import (
+from tbp.monty.frameworks.models.evidence_matching.model import (
     MontyForEvidenceGraphMatching,
 )
 from tbp.monty.frameworks.models.graph_matching import MontyForGraphMatching
