+++ src/tbp/monty/frameworks/models/evidence_matching/graph_memory.py
@@ -1,12 +1,20 @@
+# Copyright 2025 Thousand Brains Project
+#
+# Copyright may exist in Contributors' modifications
+# and/or contributions to the work.
+#
+# Use of this source code is governed by the MIT
+# license that can be found in the LICENSE file or at
+# https://opensource.org/licenses/MIT.
+
+import logging
+
 from tbp.monty.frameworks.models.graph_matching import GraphMemory
 from tbp.monty.frameworks.models.object_model import (
     GraphObjectModel,
     GridObjectModel,
     GridTooSmallError,
 )
-
-
-import logging
 
 
 class EvidenceGraphMemory(GraphMemory):
