+++ src/tbp/monty/frameworks/models/evidence_sdr_matching.py
@@ -15,7 +15,9 @@ import shutil
 import numpy as np
 from tqdm import tqdm
 
-from tbp.monty.frameworks.models.evidence_matching import EvidenceGraphLM
+from tbp.monty.frameworks.models.evidence_matching.learning_module import (
+    EvidenceGraphLM,
+)
 
 
 class LoggerSDR:
