+++ src/tbp/monty/frameworks/models/no_reset_evidence_matching.py
@@ -11,8 +11,10 @@ from typing import List
 
 import numpy as np
 
-from tbp.monty.frameworks.models.evidence_matching import (
+from tbp.monty.frameworks.models.evidence_matching.learning_module import (
     EvidenceGraphLM,
+)
+from tbp.monty.frameworks.models.evidence_matching.model import (
     MontyForEvidenceGraphMatching,
 )
 from tbp.monty.frameworks.models.mixins.no_reset_evidence import (
